requirements.txt
streamlit
python-dotenv
PyMuPDF
langchain
openai
faiss-cpu
